Crate ic_execution_environment

Structs

ExecutionEnvironmentImpl
Struct that is responsible for executing update type message messages on canisters and subnet messages.

HttpQueryHandlerImpl
Struct that is responsible for handling queries sent by user.

IngressHistoryReaderImpl
Struct that implements the ingress history reader trait. Consumers of this trait can use this to inspect the ingress history.

IngressHistoryWriterImpl
Struct that implements the ingress history writer trait. Consumers of this trait can use this to update the ingress history.

Functions

setup_execution
Helper function to constructs the public facing components that the ExecutionEnvironment crate exports.